+++ private/plugins/forum/language/french_canada_utf-8.php
@@ -941,7 +941,7 @@ $LANG_fs['forum'] = array(
 $LANG_configselects['forum'] = array(
     0 => array('True' => 1, 'False' => 0),
     1 => array('True' => true, 'False' => false),
-    2 => array('Haut de la Page' => 1, 'Histoire Après Vedette' => 2, 'Bas de la Page' => 3),
+    2 => array('Replace Home Page' => 0, 'Haut de la Page' => 1, 'Histoire Après Vedette' => 2, 'Bas de la Page' => 3),
     3 => array('Pas de Blocs' => 'noblocks', 'Gauche Blocs' => 'leftblocks', 'Blocs Droite' => 'rightblocks', 'Droite / Gauche Blocs' => 'allblocks'),
     4 => array('Bloquer Menu' => 'blockmenu', 'La Barre de Navigation' => 'navbar', 'None' => 'none'),
     5 => array('Descending' => 'DESC', 'Ascending' => 'ASC'),